policy/requirements.txt
psycopg[binary]
python-dotenv
pandas
requests
aioodbc
asyncpg
PyYAML
psycopg2-binary